refactor(ui): realign source entries for subpath exports
