+++ packages/ui/src/accordion/recipe.ts
