+++ packages/ui/src/slider/recipe.ts
