+++ packages/ui/src/stories/button.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite"
 import type { ComponentProps } from "react"
 import { expect, fn, userEvent, within } from "storybook/test"
 
-import { Button } from "../component/button/ui"
+import { Button } from "../button"
 export default {
   title: "Base/Button",
   tags: ["autodocs"],
