+++ packages/ui/src/stories/chip.stories.tsx
@@ -3,7 +3,7 @@ import { PlusCircleIcon } from "lucide-react"
 import { ComponentProps } from "react"
 import { expect, fn, userEvent } from "storybook/test"
 
-import { Chip } from "../component/chip/ui"
+import { Chip } from "../chip"
 
 export default {
   title: "Base/Chip",
