+++ packages/ui/src/stories/tabs.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite"
 import { useState } from "react"
 import { expect, userEvent, within } from "storybook/test"
 
-import * as Tabs from "../component/tabs/ui"
+import * as Tabs from "../tabs"
 
 const meta = {
   title: "Base/Tabs",
