--- packages/ui/src/types/index.ts
@@ -1,1 +1,0 @@
-export type Assign<T, U> = Omit<T, keyof U> & U
